requirements_dev.txt
pip==9.0.1
bumpversion==0.5.3
wheel==0.30.0
# watchdog==0.8.3
flake8==3.5.0
tox==2.9.1
coverage==4.5.1
# Sphinx==1.7.1
twine==1.10.0
nose==1.3.7
pyhamcrest==1.9.0
matchmock==0.7.1

# Mock is not included in Python2
mock==2.0.0; python_version=='2.7'

-r requirements.txt
